+++ game/cell.py
@@ -5,3 +5,7 @@ class Cell():
     def __init__(self, state, piece):
         self.__state__= state
         self.__piece__ = piece
+
+    def moved(self):
+        self.__state__ = True
+        self.__piece__ = None